Updates to hubremote.py
Added logging facility. Configured to use an all up log file in root of
the remotes folder. It also creates a log specific to the remote in that
remotes folder. the all up log is a consoldiation of all the individual
remote log files.

Added doc string to each of the functions in hubremote.py